Remove "bash" folder, reintroduce it as "home"
So I can store other config files I have on $HOME as well. Also, .bash_aliases has been renamed to .bash_custom, to better reflect the use of that file.

I've added a "scrub_patch" function to it that cleans patches to make them more suitable for the Gentoo tree.